+++ CaminoContents/CaminoApp.swift
@@ -1,10 +1,3 @@
-//
-//  CaminoApp.swift
-//  Camino
-//
-//  Created by Jeff White on 4/4/25.
-//
-
 import SwiftUI
 
 @main
@@ -14,4 +7,4 @@ struct CaminoApp: App {
             ContentView()
         }
     }
-}
+}